Use int64 in API messages.

+++ objects.go
@@ -3,27 +3,27 @@ package tg
 // GENERATED AUTOMATICALLY BY objects.py
 
 type Update struct {
-	UpdateID           int                 `json:"update_id"`
+	UpdateID           int64               `json:"update_id"`
 	Message            *Message            `json:"message"`
 	InlineQuery        *InlineQuery        `json:"inline_query"`
 	ChosenInlineResult *ChosenInlineResult `json:"chosen_inline_result"`
 }
 
 type GetUpdatesRequest struct {
-	Offset  *int `json:"offset,omitempty"`
-	Limit   *int `json:"limit,omitempty"`
-	Timeout *int `json:"timeout,omitempty"`
+	Offset  *int64 `json:"offset,omitempty"`
+	Limit   *int64 `json:"limit,omitempty"`
+	Timeout *int64 `json:"timeout,omitempty"`
 }
 
 type User struct {
-	ID        int     `json:"id"`
+	ID        int64   `json:"id"`
 	FirstName string  `json:"first_name"`
 	LastName  *string `json:"last_name"`
 	Username  *string `json:"username"`
 }
 
 type Chat struct {
-	ID        int     `json:"id"`
+	ID        int64   `json:"id"`
 	Type      string  `json:"type"`
 	Title     *string `json:"title"`
 	Username  *string `json:"username"`
@@ -32,12 +32,12 @@ type Chat struct {
 }
 
 type Message struct {
-	MessageID             int         `json:"message_id"`
+	MessageID             int64       `json:"message_id"`
 	From                  *User       `json:"from"`
-	Date                  int         `json:"date"`
+	Date                  int64       `json:"date"`
 	Chat                  Chat        `json:"chat"`
 	ForwardFrom           *User       `json:"forward_from"`
-	ForwardDate           *int        `json:"forward_date"`
+	ForwardDate           *int64      `json:"forward_date"`
 	ReplyToMessage        *Message    `json:"reply_to_message"`
 	Text                  *string     `json:"text"`
 	Audio                 *Audio      `json:"audio"`
@@ -57,24 +57,24 @@ type Message struct {
 	GroupChatCreated      bool        `json:"group_chat_created"`
 	SupergroupChatCreated bool        `json:"supergroup_chat_created"`
 	ChannelChatCreated    bool        `json:"channel_chat_created"`
-	MigrateToChatID       *int        `json:"migrate_to_chat_id"`
-	MigrateFromChatID     *int        `json:"migrate_from_chat_id"`
+	MigrateToChatID       *int64      `json:"migrate_to_chat_id"`
+	MigrateFromChatID     *int64      `json:"migrate_from_chat_id"`
 }
 
 type PhotoSize struct {
 	FileID   string `json:"file_id"`
-	Width    int    `json:"width"`
-	Height   int    `json:"height"`
-	FileSize *int   `json:"file_size"`
+	Width    int64  `json:"width"`
+	Height   int64  `json:"height"`
+	FileSize *int64 `json:"file_size"`
 }
 
 type Audio struct {
 	FileID    string  `json:"file_id"`
-	Duration  int     `json:"duration"`
+	Duration  int64   `json:"duration"`
 	Performer *string `json:"performer"`
 	Title     *string `json:"title"`
 	MimeType  *string `json:"mime_type"`
-	FileSize  *int    `json:"file_size"`
+	FileSize  *int64  `json:"file_size"`
 }
 
 type Document struct {
@@ -82,39 +82,39 @@ type Document struct {
 	Thumb    *PhotoSize `json:"thumb"`
 	FileName *string    `json:"file_name"`
 	MimeType *string    `json:"mime_type"`
-	FileSize *int       `json:"file_size"`
+	FileSize *int64     `json:"file_size"`
 }
 
 type Sticker struct {
 	FileID   string     `json:"file_id"`
-	Width    int        `json:"width"`
-	Height   int        `json:"height"`
+	Width    int64      `json:"width"`
+	Height   int64      `json:"height"`
 	Thumb    *PhotoSize `json:"thumb"`
-	FileSize *int       `json:"file_size"`
+	FileSize *int64     `json:"file_size"`
 }
 
 type Video struct {
 	FileID   string     `json:"file_id"`
-	Width    int        `json:"width"`
-	Height   int        `json:"height"`
-	Duration int        `json:"duration"`
+	Width    int64      `json:"width"`
+	Height   int64      `json:"height"`
+	Duration int64      `json:"duration"`
 	Thumb    *PhotoSize `json:"thumb"`
 	MimeType *string    `json:"mime_type"`
-	FileSize *int       `json:"file_size"`
+	FileSize *int64     `json:"file_size"`
 }
 
 type Voice struct {
 	FileID   string  `json:"file_id"`
-	Duration int     `json:"duration"`
+	Duration int64   `json:"duration"`
 	MimeType *string `json:"mime_type"`
-	FileSize *int    `json:"file_size"`
+	FileSize *int64  `json:"file_size"`
 }
 
 type Contact struct {
 	PhoneNumber string  `json:"phone_number"`
 	FirstName   string  `json:"first_name"`
 	LastName    *string `json:"last_name"`
-	UserID      *int    `json:"user_id"`
+	UserID      *int64  `json:"user_id"`
 }
 
 type Location struct {
@@ -123,7 +123,7 @@ type Location struct {
 }
 
 type UserProfilePhotos struct {
-	TotalCount int           `json:"total_count"`
+	TotalCount int64         `json:"total_count"`
 	Photos     [][]PhotoSize `json:"photos"`
 }
 
@@ -145,71 +145,71 @@ type ForceReply struct {
 }
 
 type SendMessageRequest struct {
-	ChatID                interface{} `json:"chat_id"` // int|string
+	ChatID                interface{} `json:"chat_id"` // int64|string
 	Text                  string      `json:"text"`
 	ParseMode             *string     `json:"parse_mode,omitempty"`
 	DisableWebPagePreview *bool       `json:"disable_web_page_preview,omitempty"`
-	ReplyToMessageID      *int        `json:"reply_to_message_id,omitempty"`
+	ReplyToMessageID      *int64      `json:"reply_to_message_id,omitempty"`
 	ReplyMarkup           interface{} `json:"reply_markup,omitempty"` // ReplyKeyboardMarkup|ReplyKeyboardHide|ForceReply
 }
 
 type ForwardMessageRequest struct {
-	ChatID     interface{} `json:"chat_id"`      // int|string
-	FromChatID interface{} `json:"from_chat_id"` // int|string
-	MessageID  int         `json:"message_id"`
+	ChatID     interface{} `json:"chat_id"`      // int64|string
+	FromChatID interface{} `json:"from_chat_id"` // int64|string
+	MessageID  int64       `json:"message_id"`
 }
 
 type SendPhotoRequest struct {
-	ChatID           interface{} `json:"chat_id"` // int|string
+	ChatID           interface{} `json:"chat_id"` // int64|string
 	Photo            interface{} `json:"photo"`   // InputFile|string
 	Caption          *string     `json:"caption,omitempty"`
-	ReplyToMessageID *int        `json:"reply_to_message_id,omitempty"`
+	ReplyToMessageID *int64      `json:"reply_to_message_id,omitempty"`
 	ReplyMarkup      interface{} `json:"reply_markup,omitempty"` // ReplyKeyboardMarkup|ReplyKeyboardHide|ForceReply
 }
 
 type SendDocumentRequest struct {
-	ChatID           interface{} `json:"chat_id"`  // int|string
+	ChatID           interface{} `json:"chat_id"`  // int64|string
 	Document         interface{} `json:"document"` // InputFile|string
-	ReplyToMessageID *int        `json:"reply_to_message_id,omitempty"`
+	ReplyToMessageID *int64      `json:"reply_to_message_id,omitempty"`
 	ReplyMarkup      interface{} `json:"reply_markup,omitempty"` // ReplyKeyboardMarkup|ReplyKeyboardHide|ForceReply
 }
 
 type SendStickerRequest struct {
-	ChatID           interface{} `json:"chat_id"` // int|string
+	ChatID           interface{} `json:"chat_id"` // int64|string
 	Sticker          interface{} `json:"sticker"` // InputFile|string
-	ReplyToMessageID *int        `json:"reply_to_message_id,omitempty"`
+	ReplyToMessageID *int64      `json:"reply_to_message_id,omitempty"`
 	ReplyMarkup      interface{} `json:"reply_markup,omitempty"` // ReplyKeyboardMarkup|ReplyKeyboardHide|ForceReply
 }
 
 type SendVideoRequest struct {
-	ChatID           interface{} `json:"chat_id"` // int|string
+	ChatID           interface{} `json:"chat_id"` // int64|string
 	Video            interface{} `json:"video"`   // InputFile|string
-	Duration         *int        `json:"duration,omitempty"`
+	Duration         *int64      `json:"duration,omitempty"`
 	Caption          *string     `json:"caption,omitempty"`
-	ReplyToMessageID *int        `json:"reply_to_message_id,omitempty"`
+	ReplyToMessageID *int64      `json:"reply_to_message_id,omitempty"`
 	ReplyMarkup      interface{} `json:"reply_markup,omitempty"` // ReplyKeyboardMarkup|ReplyKeyboardHide|ForceReply
 }
 
 type SendVoiceRequest struct {
-	ChatID           interface{} `json:"chat_id"` // int|string
+	ChatID           interface{} `json:"chat_id"` // int64|string
 	Voice            interface{} `json:"voice"`   // InputFile|string
-	Duration         *int        `json:"duration,omitempty"`
-	ReplyToMessageID *int        `json:"reply_to_message_id,omitempty"`
+	Duration         *int64      `json:"duration,omitempty"`
+	ReplyToMessageID *int64      `json:"reply_to_message_id,omitempty"`
 	ReplyMarkup      interface{} `json:"reply_markup,omitempty"` // ReplyKeyboardMarkup|ReplyKeyboardHide|ForceReply
 }
 
 type SendLocationRequest struct {
-	ChatID           interface{} `json:"chat_id"` // int|string
+	ChatID           interface{} `json:"chat_id"` // int64|string
 	Latitude         float64     `json:"latitude"`
 	Longitude        float64     `json:"longitude"`
-	ReplyToMessageID *int        `json:"reply_to_message_id,omitempty"`
+	ReplyToMessageID *int64      `json:"reply_to_message_id,omitempty"`
 	ReplyMarkup      interface{} `json:"reply_markup,omitempty"` // ReplyKeyboardMarkup|ReplyKeyboardHide|ForceReply
 }
 
 type GetUserProfilePhotosRequest struct {
-	UserID int  `json:"user_id"`
-	Offset *int `json:"offset,omitempty"`
-	Limit  *int `json:"limit,omitempty"`
+	UserID int64  `json:"user_id"`
+	Offset *int64 `json:"offset,omitempty"`
+	Limit  *int64 `json:"limit,omitempty"`
 }
 
 type GetFileRequest struct {
@@ -226,7 +226,7 @@ type InlineQuery struct {
 type AnswerInlineQueryRequest struct {
 	InlineQueryID string              `json:"inline_query_id"`
 	Results       []InlineQueryResult `json:"results"`
-	CacheTime     *int                `json:"cache_time,omitempty"`
+	CacheTime     *int64              `json:"cache_time,omitempty"`
 	IsPersonal    *bool               `json:"is_personal,omitempty"`
 	NextOffset    *string             `json:"next_offset,omitempty"`
 }
@@ -244,16 +244,16 @@ type InlineQueryResultArticle struct {
 	HideURL               *bool   `json:"hide_url"`
 	Description           *string `json:"description"`
 	ThumbURL              *string `json:"thumb_url"`
-	ThumbWidth            *int    `json:"thumb_width"`
-	ThumbHeight           *int    `json:"thumb_height"`
+	ThumbWidth            *int64  `json:"thumb_width"`
+	ThumbHeight           *int64  `json:"thumb_height"`
 }
 
 type InlineQueryResultPhoto struct {
 	Type                  string  `json:"type"`
 	ID                    string  `json:"id"`
 	PhotoURL              string  `json:"photo_url"`
-	PhotoWidth            *int    `json:"photo_width"`
-	PhotoHeight           *int    `json:"photo_height"`
+	PhotoWidth            *int64  `json:"photo_width"`
+	PhotoHeight           *int64  `json:"photo_height"`
 	ThumbURL              string  `json:"thumb_url"`
 	Title                 *string `json:"title"`
 	Description           *string `json:"description"`
@@ -267,8 +267,8 @@ type InlineQueryResultGif struct {
 	Type                  string  `json:"type"`
 	ID                    string  `json:"id"`
 	GIFURL                string  `json:"gif_url"`
-	GIFWidth              *int    `json:"gif_width"`
-	GIFHeight             *int    `json:"gif_height"`
+	GIFWidth              *int64  `json:"gif_width"`
+	GIFHeight             *int64  `json:"gif_height"`
 	ThumbURL              string  `json:"thumb_url"`
 	Title                 *string `json:"title"`
 	Caption               *string `json:"caption"`
@@ -281,8 +281,8 @@ type InlineQueryResultMpeg4Gif struct {
 	Type                  string  `json:"type"`
 	ID                    string  `json:"id"`
 	MPEG4URL              string  `json:"mpeg4_url"`
-	MPEG4Width            *int    `json:"mpeg4_width"`
-	MPEG4Height           *int    `json:"mpeg4_height"`
+	MPEG4Width            *int64  `json:"mpeg4_width"`
+	MPEG4Height           *int64  `json:"mpeg4_height"`
 	ThumbURL              string  `json:"thumb_url"`
 	Title                 *string `json:"title"`
 	Caption               *string `json:"caption"`
@@ -299,9 +299,9 @@ type InlineQueryResultVideo struct {
 	MessageText           string  `json:"message_text"`
 	ParseMode             *string `json:"parse_mode"`
 	DisableWebPagePreview *bool   `json:"disable_web_page_preview"`
-	VideoWidth            *int    `json:"video_width"`
-	VideoHeight           *int    `json:"video_height"`
-	VideoDuration         *int    `json:"video_duration"`
+	VideoWidth            *int64  `json:"video_width"`
+	VideoHeight           *int64  `json:"video_height"`
+	VideoDuration         *int64  `json:"video_duration"`
 	ThumbURL              string  `json:"thumb_url"`
 	Title                 string  `json:"title"`
 	Description           *string `json:"description"`
